requirements.txt
playwright>=1.40.0
pywinauto
Appium-Python-Client
